refactor: HttpHeaderType을 적절한 패키지로 이동
